feat(blackjack) : DIVIDE - Score class and Winner object
