+++ src/main/kotlin/README.md
@@ -23,6 +23,6 @@
   - [X] 처음에 받은 두 장의 합계가 16 이하이면 반드시 카드 하나를 추가로 뽑아야 함
   - [X] 17점 이상이면 추가로 받을 수 없음
   - [ ] 처음에 받은 카드를 출력할 때, 딜러는 한 장을 출력하지 않음
-  - [ ] 21을 초과하면 남아있는 플레이어 우승
+  - [X] 21을 초과하면 남아있는 플레이어 우승
 - [ ] 승패 출력
   - 딜러의 점수를 기준으로 승패 출력
