+++ src/main/kotlin/blackjack/model/Score.kt
@@ -1,0 +1,3 @@
+package blackjack.model
+
+data class Score(val win: Int, val lose: Int)
