+++ src/test/kotlin/blackjack/model/WinnerTest.kt
@@ -31,23 +31,3 @@ class WinnerTest {
         assertThat(totalScore.first().lose).isEqualTo(1)
     }
 }
-
-object Winner {
-
-    fun getTotalScore(dealer: Gamer, players: List<Gamer>): List<Score> {
-        val playersWinOrNot = players
-            .map { it.totalPoints }
-            .map {
-                if (it > Gamer.MAX_POINT) return@map false
-                if (dealer.totalPoints > Gamer.MAX_POINT) return@map true
-                return@map dealer.totalPoints < it
-            }
-        val dealerScore = Score(win = playersWinOrNot.count { !it }, lose = playersWinOrNot.count { it })
-        val playersScore = playersWinOrNot.map {
-            if (it) Score(win = 1, lose = 0) else Score(win = 0, lose = 1)
-        }
-        return listOf(listOf(dealerScore), playersScore).flatten()
-    }
-}
-
-data class Score(val win: Int, val lose: Int)
